Move typings to devDependencies
Leaving @types in dependencies causes them to be installed when @easy-webpack/core is installed, inferring with the application using easy-webpack. Typings are not required for using easy-webpack, only for development of easy-webpack. It's especially important in case of @types/nodes which overrides some methods such as setTimeout which causes compilation error.